+++ src/main/java/org/songeun/petdongne_server/survey/domain/SurveyOptionStat.java
@@ -1,11 +1,11 @@
-package org.songeun.petdongne_server.survey.entity;
+package org.songeun.petdongne_server.survey.domain;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
-import org.songeun.petdongne_server.residentialComplex.entity.ResidentialComplex;
+import org.songeun.petdongne_server.residentialComplex.domain.ResidentialComplex;
 import org.songeun.petdongne_server.global.common.BaseEntity;
 
 @Entity
